+++ packages/elements/emoji/src/__tests__/unit/components/picker/EmojiPickerListSpec.tsx
@@ -1,26 +1,26 @@
-import { mount, ReactWrapper } from 'enzyme';
-import * as React from 'react';
+import CrossCircleIcon from '@atlaskit/icon/glyph/cross-circle';
 import { expect } from 'chai';
-
+import { ReactWrapper } from 'enzyme';
+import { mountWithIntl } from '@atlaskit/editor-test-helpers';
+import * as React from 'react';
 import { List as VirtualList } from 'react-virtualized/dist/commonjs/List';
-import * as styles from '../../../../components/picker/styles';
-import EmojiPickerList from '../../../../components/picker/EmojiPickerList';
-import EmojiPickerCategoryHeading from '../../../../components/picker/EmojiPickerCategoryHeading';
-import {
-  imageEmoji,
-  siteEmojiFoo,
-  siteEmojiWtf,
-  emojis as allEmojis,
-  onRowsRenderedArgs,
-  atlassianEmojis,
-} from '../../_test-data';
-import { EmojiDescription } from '../../../../types';
-import { deleteEmojiLabel } from '../../../../constants';
 import {
   CachingEmoji,
   CachingEmojiProps,
 } from '../../../../components/common/CachingEmoji';
-import CrossCircleIcon from '@atlaskit/icon/glyph/cross-circle';
+import EmojiPickerCategoryHeading from '../../../../components/picker/EmojiPickerCategoryHeading';
+import EmojiPickerList from '../../../../components/picker/EmojiPickerList';
+import * as styles from '../../../../components/picker/styles';
+import { deleteEmojiLabel } from '../../../../constants';
+import { EmojiDescription } from '../../../../types';
+import {
+  atlassianEmojis,
+  emojis as allEmojis,
+  imageEmoji,
+  onRowsRenderedArgs,
+  siteEmojiFoo,
+  siteEmojiWtf,
+} from '../../_test-data';
 
 const emojis = [imageEmoji];
 const customEmojis: EmojiDescription[] = [siteEmojiFoo, siteEmojiWtf];
@@ -28,13 +28,13 @@ const customEmojis: EmojiDescription[] = [siteEmojiFoo, siteEmojiWtf];
 describe('<EmojiPickerList />', () => {
   describe('list', () => {
     it('should contain search ', () => {
-      const wrapper = mount(<EmojiPickerList emojis={emojis} />);
+      const wrapper = mountWithIntl(<EmojiPickerList emojis={emojis} />);
 
       expect(wrapper.find(`.${styles.pickerSearch}`)).to.have.length(1);
     });
 
     it('should show people category first if no frequently used', () => {
-      const wrapper = mount(<EmojiPickerList emojis={emojis} />);
+      const wrapper = mountWithIntl(<EmojiPickerList emojis={emojis} />);
 
       const categoryHeadings = wrapper.find(EmojiPickerCategoryHeading);
       expect(categoryHeadings.get(0).props.id).to.equal('PEOPLE');
@@ -59,7 +59,9 @@ describe('<EmojiPickerList />', () => {
 
       const emojisWithFrequent = [...emojis, frequentEmoji];
 
-      const wrapper = mount(<EmojiPickerList emojis={emojisWithFrequent} />);
+      const wrapper = mountWithIntl(
+        <EmojiPickerList emojis={emojisWithFrequent} />,
+      );
 
       const categoryHeadings = wrapper.find(EmojiPickerCategoryHeading);
       expect(categoryHeadings.get(0).props.id).to.equal('FREQUENT');
@@ -77,7 +79,9 @@ describe('<EmojiPickerList />', () => {
           order: 0,
         },
       ];
-      const wrapper = mount(<EmojiPickerList emojis={outOfOrderEmojis} />);
+      const wrapper = mountWithIntl(
+        <EmojiPickerList emojis={outOfOrderEmojis} />,
+      );
 
       const cachingEmojis: ReactWrapper<
         CachingEmojiProps,
@@ -107,7 +111,7 @@ describe('<EmojiPickerList />', () => {
         },
       ];
 
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={frequentCategoryEmojis} />,
       );
 
@@ -128,14 +132,18 @@ describe('<EmojiPickerList />', () => {
 
   describe('custom upload display', () => {
     it('should render user custom emojis under Your Uploads', () => {
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={customEmojis} currentUser={{ id: 'hulk' }} />,
       );
 
       const categoryHeadings = wrapper.find(EmojiPickerCategoryHeading);
       expect(categoryHeadings.length).to.equal(2);
-      expect(categoryHeadings.get(0).props.title).to.equal('Your uploads');
-      expect(categoryHeadings.get(1).props.title).to.equal('All uploads');
+      expect(categoryHeadings.get(0).props.title).to.equal(
+        'userUploadsCustomCategory',
+      );
+      expect(categoryHeadings.get(1).props.title).to.equal(
+        'allUploadsCustomCategory',
+      );
 
       const cachedEmojis = wrapper.find(CachingEmoji);
 
@@ -147,13 +155,15 @@ describe('<EmojiPickerList />', () => {
     });
 
     it('should not render user custom emojis section if user has none', () => {
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={customEmojis} currentUser={{ id: 'alex' }} />,
       );
 
       const categoryHeadings = wrapper.find(EmojiPickerCategoryHeading);
       expect(categoryHeadings.length).to.equal(1);
-      expect(categoryHeadings.get(0).props.title).to.equal('All uploads');
+      expect(categoryHeadings.get(0).props.title).to.equal(
+        'allUploadsCustomCategory',
+      );
 
       const cachedEmojis = wrapper.find(CachingEmoji);
 
@@ -163,11 +173,13 @@ describe('<EmojiPickerList />', () => {
     });
 
     it('should not render user custom emojis section if currentUser is undefined', () => {
-      const wrapper = mount(<EmojiPickerList emojis={customEmojis} />);
+      const wrapper = mountWithIntl(<EmojiPickerList emojis={customEmojis} />);
 
       const categoryHeadings = wrapper.find(EmojiPickerCategoryHeading);
       expect(categoryHeadings.length).to.equal(1);
-      expect(categoryHeadings.get(0).props.title).to.equal('All uploads');
+      expect(categoryHeadings.get(0).props.title).to.equal(
+        'allUploadsCustomCategory',
+      );
 
       const cachedEmojis = wrapper.find(CachingEmoji);
 
@@ -178,7 +190,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should trigger onCategoryActivated', () => {
       const onCategoryActivated = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={allEmojis}
           onCategoryActivated={onCategoryActivated}
@@ -199,7 +211,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should not break while finding category in an empty list', () => {
       const onCategoryActivated = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={[]}
           onCategoryActivated={onCategoryActivated}
@@ -219,7 +231,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should trigger onCategoryActivated for first category', () => {
       const onCategoryActivated = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={allEmojis}
           onCategoryActivated={onCategoryActivated}
@@ -240,7 +252,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should trigger onCategoryActivated for bottom category', () => {
       const onCategoryActivated = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={allEmojis}
           onCategoryActivated={onCategoryActivated}
@@ -261,7 +273,7 @@ describe('<EmojiPickerList />', () => {
 
   describe('delete', () => {
     it('should render user custom emoji with delete button', () => {
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={customEmojis} currentUser={{ id: 'hulk' }} />,
       );
       const yourEmoji = wrapper.find(CachingEmoji).at(0);
@@ -271,7 +283,7 @@ describe('<EmojiPickerList />', () => {
     });
 
     it('should not render delete button if not user custom emoji', () => {
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={customEmojis} currentUser={{ id: 'alex' }} />,
       );
       const emoji = wrapper.find(CachingEmoji).at(0);
@@ -281,7 +293,7 @@ describe('<EmojiPickerList />', () => {
     });
 
     it('should have label "delete-emoji" on delete button', () => {
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList emojis={customEmojis} currentUser={{ id: 'hulk' }} />,
       );
       const deleteButton = wrapper
@@ -297,7 +309,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should call onEmojiDelete if delete button is clicked', () => {
       const onDelete = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={customEmojis}
           currentUser={{ id: 'hulk' }}
@@ -314,7 +326,7 @@ describe('<EmojiPickerList />', () => {
 
     it('should not call onEmojiSelected if delete button is clicked', () => {
       const onSelection = jest.fn();
-      const wrapper = mount(
+      const wrapper = mountWithIntl(
         <EmojiPickerList
           emojis={customEmojis}
           currentUser={{ id: 'hulk' }}