+++ packages/elements/emoji/src/components/common/styles.ts
@@ -449,8 +449,12 @@ export const emojiPreviewErrorMessage = style({
   alignItems: 'center',
 });
 
+export const addCustomEmojiButton = style({
+  maxWidth: '285px',
+});
+
 export const uploadRetryButton = style({
-  width: '93px',
+  maxWidth: '172px',
   justifyContent: 'center',
   fontWeight: 'bold',
   marginRight: '4px',
@@ -462,7 +466,7 @@ export const uploadRetryButton = style({
 });
 
 export const uploadEmojiButton = style({
-  width: '93px',
+  maxWidth: '187px',
   justifyContent: 'center',
   marginRight: '4px',
   $nest: {
@@ -470,4 +474,13 @@ export const uploadEmojiButton = style({
       display: 'flex',
     },
   },
+});
+
+export const cancelButton = style({
+  maxWidth: '100px',
+});
+
+export const buttonSpinner = style({
+  marginRight: '10px',
+  marginLeft: '10px',
 });